Remove NaN's from sythetic probabilities
When sampling from a data set, the resampled points can sometimes be invalid for a specific model, which results in `nan` values on the computed probability. This causes an error when trying to fit the linear model to those probabilities.

This PR remove the points that produce invalid probabilities, and delays computing the distances until said points have been removed.